+++ asf_search/search/search.py
@@ -42,6 +42,8 @@ def search(
         temporalBaselineDays: Union[str, Sequence[str]] = None,
         operaBurstID: Union[str, Sequence[str]] = None,
         dataset: Union[str, Sequence[str]] = None,
+        shortName: Union[str, Sequence[str]] = None,
+        cmr_keywords: Union[Tuple[str, str], Sequence[Tuple[str, str]]] = None,
         maxResults: int = None,
         opts: ASFSearchOptions = None,
 ) -> ASFSearchResults: