+++ asf_search/search/search_count.py
@@ -44,6 +44,8 @@ def search_count(
         temporalBaselineDays: Union[str, Sequence[str]] = None,
         operaBurstID: Union[str, Sequence[str]] = None,
         dataset: Union[str, Sequence[str]] = None,
+        shortName: Union[str, Sequence[str]] = None,
+        cmr_keywords: Union[Tuple[str, str], Sequence[Tuple[str, str]]] = None,
         maxResults: int = None,
         opts: ASFSearchOptions = None,
 ) -> int: